requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.12
langchain-community>=0.0.28
faiss-cpu==1.7.4
tiktoken==0.6.0 
tiktoken==0.6.0 